split login page into 2 pages: login and register
