+++ backend/DestionationBucketListBackend/DestionationBucketListBackend/Program.cs
@@ -23,9 +23,7 @@ if (app.Environment.IsDevelopment())
     app.UseSwagger();
     app.UseSwaggerUI();
 }
-{
-    
-}
+
 app.UseAuthorization();
 app.MapControllers();
 app.Run();
