+++ frontend/src/Pages/Register/Register.tsx
@@ -1,65 +1,13 @@
-import "bootstrap/dist/css/bootstrap.min.css";
 import React, { useState } from "react";
-import { User } from "./Model/User";
+import { User } from "../../Model/User";
+import { useNavigate } from "react-router-dom";
 
-export default function () {
-  const [authMode, setLoginMode] = useState("signin");
+export const Register = () => {
   const [user, setUser] = useState<User>();
   const [name, setName] = useState<string>("");
   const [email, setEmail] = useState<string>("");
   const [password, setPassword] = useState<string>("");
-
-  const changeLoginMode = () => {
-    setLoginMode(authMode === "signin" ? "signup" : "signin");
-  };
-
-  if (authMode === "signin") {
-    return (
-      <div className="Login-form-container">
-        <form className="Login-form">
-          <div className="Login-form-content">
-            <h3 className="Login-form-title">Sign In</h3>
-            <div className="text-center">
-              Not registered yet?{" "}
-              <span className="link-primary" onClick={changeLoginMode}>
-                Sign Up
-              </span>
-            </div>
-            <div className="form-group mt-3">
-              <label>Email address</label>
-              <input
-                type="email"
-                className="form-control mt-1"
-                placeholder="Enter email"
-                onChange={(x) => {
-                  setEmail(x.currentTarget.innerText);
-                }}
-              />
-            </div>
-            <div className="form-group mt-3">
-              <label>Password</label>
-              <input
-                type="password"
-                className="form-control mt-1"
-                placeholder="Enter password"
-                onChange={(x) => {
-                  setPassword(x.currentTarget.innerText);
-                }}
-              />
-            </div>
-            <div className="d-grid gap-2 mt-3">
-              <button type="submit" className="submit-btn btn btn-primary">
-                Submit
-              </button>
-            </div>
-            <p className="text-center mt-2">
-              <a href="#">Forgot password?</a>
-            </p>
-          </div>
-        </form>
-      </div>
-    );
-  }
+  const navigate = useNavigate();
 
   return (
     <div className="Login-form-container">
@@ -68,7 +16,14 @@ export default function () {
           <h3 className="Login-form-title">Create an account</h3>
           <div className="text-center">
             Already registered?{" "}
-            <span className="link-primary" onClick={changeLoginMode}>
+            <span
+              className="link-primary"
+              onClick={() => {
+                navigate("/login", {
+                  replace: true,
+                });
+              }}
+            >
               Sign In
             </span>
           </div>
@@ -127,7 +82,6 @@ export default function () {
                 };
                 let response = await fetch(url, header);
                 console.log(response);
-                //await response;
               }}
             >
               Submit
@@ -137,4 +91,4 @@ export default function () {
       </form>
     </div>
   );
-}
+};
